# catalog-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for a catalog service.",
        "title": "Catalog Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3003",
    "basePath": "/",
    "paths": {
...
        "/api/v1/catalog/choices/{group_id}": {
            "delete": {
                "description": "Delete a BOM choice group by its ID",
                "tags": [
                    "choices"
                ],
                "summary": "Delete a choice group",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Group ID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/catalog/choices/{group_id}/options": {
            "post": {
                "description": "Add an ingredient option to a BOM choice group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "choices"
                ],
                "summary": "Add a choice option",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Group ID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Option Data",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.addChoiceOptionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.BOMChoiceOption"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/catalog/menus/{id}/choices": {
            "get": {
                "description": "Get the Bill of Materials choice groups for a specific menu item",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "choices"
                ],
                "summary": "Get choices by Menu Item ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/model.BOMChoiceGroup"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new BOM choice group for a menu item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "choices"
                ],
                "summary": "Create a choice group",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Choice Group Data",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BOMChoiceGroup"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.BOMChoiceGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/v1/catalog/options/{option_id}": {
            "delete": {
                "description": "Delete a choice option by its ID",
                "tags": [
                    "choices"
                ],
                "summary": "Delete a choice option",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Option ID",
                        "name": "option_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.addChoiceOptionRequest": {
            "type": "object",
            "properties": {
                "extra_price": {
                    "type": "number"
                },
                "ingredient_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                }
            }
        },
...
        "model.BOMChoiceGroup": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_required": {
                    "description": "must pick one",
                    "type": "boolean"
                },
                "max_choices": {
                    "type": "integer"
                },
                "menu_item_id": {
                    "type": "integer"
                },
                "min_choices": {
                    "type": "integer"
                },
                "name": {
                    "description": "e.g. \"เลือกเส้น\"",
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BOMChoiceOption"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.BOMChoiceOption": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "extra_price": {
                    "type": "number"
                },
                "group_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "ingredient": {
                    "$ref": "#/definitions/model.Ingredient"
                },
                "ingredient_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.Ingredient": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "unit": {
                    "description": "e.g. \"g\", \"ml\", \"piece\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}